+++ src/main/kotlin/pl/vemu/diamold/blocks/DiamoldBlock.kt
@@ -9,7 +9,7 @@ import pl.vemu.diamold.items.JsonName
 
 object DiamoldBlock : Block(
     FabricBlockSettings.of(Material.METAL).requiresTool().strength(5.0F, 6.0F)
-        .sounds(BlockSoundGroup.METAL).breakByTool(FabricToolTags.PICKAXES, 3)
+        .sounds(BlockSoundGroup.METAL).breakByTool(FabricToolTags.PICKAXES, 2)
 ), JsonName {
     override val jsonName = "diamold_block"
 }